+++ packages/expo-ui/android/src/main/java/expo/modules/ui/ExpoUIModule.kt
@@ -64,6 +64,8 @@ class ExpoUIModule : Module() {
 
     View(RNHostView::class)
 
+    View(LoadingView::class)
+
     View(TextInputView::class) {
       Events("onValueChanged")
       Prop("defaultValue", "") { view: TextInputView, text: String ->