+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -34,6 +34,7 @@ export * from './PullToRefreshBox';
 export * from './RadioButton';
 export * from './Surface';
 export * from './Text';
+export * from './LoadingIndicator';
 
 export * from './AnimatedVisibility';
 export * from './Box';